0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -24,7 +24,7 @@ for (i = 0; i < size; i++)
 printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 if (array[i] == value)
 {
-return (int)(i);
+return (i);
 }
 }
